refactor: app コードの整理・最適化とReact辞書の改善
- app/types.ts を app/dictionary.ts に統合し、型定義とデータを一元管理
- 非同期ローディングを同期処理に変更 (import.meta.glob eager loading)
- カスタムフック useDictionaryState で状態管理ロジックを抽出
- VCF生成を小さな関数に分割し、可読性向上
- 不要なコメントとActionButtonsコンポーネントを削除
- ダウンロードファイル名に日時追加、VCFメモ欄にURL追加
- React辞書の読みをひらがなからカタカナに変更 (音声認識精度向上)
- 単語リストの偶数奇数行に背景色を追加、ランダム表示の改善
- フッターに貢献歓迎メッセージとGitHubリンクを追加
