+++ app/vcf.ts
@@ -1,38 +1,48 @@
-import type { Dictionary, DictionaryEntry } from "./types";
+import type { Dictionary, DictionaryEntry } from "./dictionary";
 
-export function generateVCF(dictionaries: Dictionary[]): string {
+const GITHUB_PAGES_URL = "https://pokutuna.github.io/dictionary.vcf/";
+
+function getSelectedEntries(dictionaries: Dictionary[]): DictionaryEntry[] {
   const selectedEntries: DictionaryEntry[] = [];
-
   dictionaries.forEach((dict) => {
     dict.entries.forEach((entry) => {
-      // Include all selected entries (both individually selected and from selected dictionaries)
       if (entry.selected) {
         selectedEntries.push(entry);
       }
     });
   });
+  return selectedEntries;
+}
 
-  // Remove duplicates by word and sort
-  const uniqueEntries = Array.from(
-    new Map(selectedEntries.map((entry) => [entry.word, entry])).values()
+function deduplicateEntries(entries: DictionaryEntry[]): DictionaryEntry[] {
+  return Array.from(
+    new Map(entries.map((entry) => [entry.word, entry])).values()
   ).sort((a, b) => a.word.localeCompare(b.word));
+}
 
-  const vcfContent = uniqueEntries
-    .map(
-      (entry) =>
-        `BEGIN:VCARD\nVERSION:3.0\nFN:${entry.word}\nX-PHONETIC-LAST-NAME:${entry.reading}\nEND:VCARD`
-    )
+function createVCardString(entry: DictionaryEntry): string {
+  return `BEGIN:VCARD\nVERSION:3.0\nFN:${entry.word}\nX-PHONETIC-LAST-NAME:${entry.reading}\nNOTE:${GITHUB_PAGES_URL}\nEND:VCARD`;
+}
+
+export function generateVCF(dictionaries: Dictionary[]): string {
+  const selectedEntries = getSelectedEntries(dictionaries);
+  const uniqueEntries = deduplicateEntries(selectedEntries);
+  
+  return uniqueEntries
+    .map(createVCardString)
     .join("\n\n");
-
-  return vcfContent;
 }
 
 export function downloadVCF(content: string) {
+  const now = new Date();
+  const timestamp = now.toISOString().slice(0, 19).replace(/[:-]/g, '');
+  const filename = `dictionary_${timestamp}.vcf`;
+  
   const blob = new Blob([content], { type: "text/vcard" });
   const url = URL.createObjectURL(blob);
   const a = document.createElement("a");
   a.href = url;
-  a.download = "dictionary.vcf";
+  a.download = filename;
   document.body.appendChild(a);
   a.click();
   document.body.removeChild(a);
