+++ dictionaries/lib-react.csv
@@ -1,22 +1,22 @@
-children,ちるどれん
-Component,こんぽーねんと
-Error Boundary,えらーばうんだり
-hooks,ふっくす
-jsx,じぇーえすえっくす
-npx,えぬぴーえっくす
-props,ぷろっぷす
-React,りあくと
-ReactDOM,りあくとどむ
-reducer,りでゅーさー
-Suspence,さすぺんす
-tsx,てぃーえすえっくす
-useCallback,ゆーずこーるばっく
-useContext,ゆーずこんてきすと
-useEffect,ゆーずえふぇくと
-useLayoutEffect,ゆーずれいあうとえふぇくと
-useMemo,ゆーずめも
-useQuery,ゆーずくえりー
-useReducer,ゆーずりでゅーさー
-useRef,ゆーずれふ
-useState,ゆーずすてーと
-useTransition,ゆーずとらんじしょん
+children,チルドレン
+Component,コンポーネント
+Error Boundary,エラーバウンダリ
+hooks,フックス
+jsx,ジェーエスエックス
+npx,エヌピーエックス
+props,プロップス
+React,リアクト
+ReactDOM,リアクトドム
+reducer,リデューサー
+Suspence,サスペンス
+tsx,ティーエスエックス
+useCallback,ユーズコールバック
+useContext,ユーズコンテキスト
+useEffect,ユーズエフェクト
+useLayoutEffect,ユーズレイアウトエフェクト
+useMemo,ユーズメモ
+useQuery,ユーズクエリー
+useReducer,ユーズリデューサー
+useRef,ユーズレフ
+useState,ユーズステート
+useTransition,ユーズトランジション
