+++ docs/tasks/005-improve-app.md
@@ -1,0 +1,43 @@
+作成したアプリケーションに細かな改善と修正を加えていきます
+
+## コードの整理・移動
+以下の方針でコードを整理・移動してください
+- 主な型定義は app/types.ts に配置
+  - dictionaries.tsx の各種型を app/types.ts へ移動
+- vcf ファイル作成に関する実装や型は app/vcf.ts に配置
+- React コンポーネントは app/components.tsx に配置
+  - 以下のようなコンポーネントが切り出せる
+  - 1つの辞書表示コンポーネント
+  - 単語リスト管理コンポーネント
+  - 単語コンポーネント
+
+## アプリケーションの改修
+- 日本語辞書なので日本語でページを作成
+- ページを1つにまとめてください
+  - トップページ home.tsx の内容を小さくする
+    - プロジェクトの説明について上部に短い領域を置くていど
+    - GitHub リポジトリへのリンクを追加
+  - dictionary.tsx の内容をトップページへ移動
+- 最初からすべての辞書をページロードされているようにする
+  - たんに import していればよい
+- 読み仮名をデフォルトから修正した場合は、単語コンポーネントが強調表示されるように
+- 表示、見せ方
+  - 辞書、単語選択の縦幅をより小さく
+  - 開閉ボタンや選択状態切り替えボタンはテキストではなくアイコンに
+  - select all, clear all は不要(辞書自体のチェックボックスで代用可能なため)
+  - lucide-react を用いて UI にわかりやすいアイコンを追加
+
+---
+
+- 辞書の表示を更新しましょう
+  - decription は要りません
+  - 単語数は辞書の名前のすぐ横にして1行に
+  - 辞書名のところにランダムに単語を選択して10個程度表示、overflow するものは ... で省略する
+  - チェックボックスの領域を少し広めに、チェックボックス以外のリストをクリックしたら開閉するように
+  - 編集済みの強調は、単語全体ではなく読み仮名入力欄の強調にする、またわかりやすいようにアイコンを表示する
+
+---
+
+辞書同士を区切るセクションを、1行にしてください\
+**プログラミング言語** 言語固有の用語とフレームワーク\
+のようなイメージ、サブタイトルの今のスタイルはいい感じなので維持して
